requirements.txt
requests
beautifulsoup4
lxml
PyYAML
colorama